+++ visualization/app/codeCharta/state/store/dynamicSettings/margin/margin.actions.ts
@@ -18,4 +18,4 @@ export function setMargin(margin: number = defaultMargin): SetMarginAction {
 	}
 }
 
-export const defaultMargin = null
+export const defaultMargin = 50